requirements.txt
numpy 
pandas
matplotlib
opencv-python
tensorflow
keras
flask
flask_mysqld